Initial refactoring of the code.
Changes:
- Moved files into a more organized structure.
- Refactored Final and Intro components from classes to functions.

Dev Tests: App was not functional previously. Ensured that it is not more broken by these changes.
