+++ src/containers/Start/Start.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
-import SignUp from './SignUp';
-import Login from './Login';
-import logo from './welcome.png'
+import SignUp from '../SignUp/SignUp';
+import Login from '../Login/Login';
+import logo from '../../assets/welcome.png'
 
 class Start extends React.Component {
   constructor(props) {
