+++ src/helpers/registerServiceWorker.js
