+++ src/index.js
@@ -1,11 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import App from './containers/App';
-import Start from './Start';
-import registerServiceWorker from './registerServiceWorker';
-import './welcome.png';
-import Control from './Control';
+import registerServiceWorker from './helpers/registerServiceWorker';
+import './assets/welcome.png';
+import Control from './containers/Control/Control';
 
 ReactDOM.render(<Control/>, document.getElementById('root'));
